Fix: null-ls completer not resolving futures properly
Possibly something upstream in plenary changed, but the way we were
creating lists made `await_all` treat them as empty, it seems. This
introduces a new `futures_map` helper that simplifies mapping from a
list of things to a list of futures *correctly*, which we do in a few
places.