Fix: nvim "enhanced" PR editor not working on latest nightlies
Not sure why I was using `NVIM_LISTEN_ADDRESS` because that's definitely
wrong. That's equivalent to `--listen`, and we want to *talk* to that
address. It [seems
like](https://github.com/neovim/neovim/issues/30256#issuecomment-2328506491)
the server address stuff was just silently ignored before, where it's
now validated. So, on older versions, it's effectively just like any old
env var (since the address will always be in use) while on the latest
nightlies this condition is (rightly) a fatal error.

We now use a different env var to avoid this conflict, and everything is
peachy again.